% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweet_functions.R
\name{ageRangeSWEET}
\alias{ageRangeSWEET}
\title{Classify a numeric column of ages using SWEET ages classification}
\usage{
ageRangeSWEET(ages, pediatric = FALSE, factors = TRUE)
}
\arguments{
\item{ages}{vector of ages}

\item{pediatric}{if TRUE, marks non-pediatric patients as Adults}

\item{factors}{if TRUE, return vector of factors. Default: TRUE}
}
\value{
SWEET age ranges vector
}
\description{
Classify a numeric column of ages using SWEET ages classification
}
\author{
Fernando Moreno Jabato <fmjabato(at)gmail(dot)com>
}
