% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{allowedAndiacareReportCols}
\alias{allowedAndiacareReportCols}
\title{Return list of allowed columns for reports}
\usage{
allowedAndiacareReportCols()
}
\value{
vector of allowed columnames
}
\description{
Return list of allowed columns for reports
}
\author{
Fernando Moreno Jabato <fmjabato(at)gmail(dot)com>
}
