% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_functions.R
\name{allowedDiabetesClassifiers}
\alias{allowedDiabetesClassifiers}
\title{Returns allowed classification algorithms}
\usage{
allowedDiabetesClassifiers()
}
\value{
list with allowed algorithms and its classifiers
}
\description{
Returns allowed classification algorithms
}
\author{
Fernando Moreno Jabato <jabato(at)uma(dot)com
}
