% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_andiacare.R
\name{andiacare}
\alias{andiacare}
\title{Classifies a dataframe of patients using Andiacare algorithm}
\usage{
andiacare(df, TIRcol = "TIR", TBRcol = "TBR", TARcol = "TAR")
}
\arguments{
\item{df}{of patients to be classified}

\item{TIRcol}{TIR info column name}

\item{TBRcol}{TBR info column name}

\item{TARcol}{TAR info column name}
}
\value{
vector with patients classes
}
\description{
Classifies a dataframe of patients using Andiacare algorithm
}
\author{
Fernando Moreno Jabato <fmjabato(at)gmail(dot)com>
}
