% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attd19_functions.R
\name{andiacareATTDTargets}
\alias{andiacareATTDTargets}
\title{Check ATTD19 targets and returns a table with checks}
\usage{
andiacareATTDTargets(df)
}
\arguments{
\item{df}{to be checked (must contain normalized names)}
}
\value{
dataframe with checks
}
\description{
Check ATTD19 targets and returns a table with checks
}
\author{
Fernando Moreno Jabato <fmjabato(at)gmail(dot)com>
}
