% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_andiacare.R
\name{andiacareClassifiers}
\alias{andiacareClassifiers}
\title{Returns, ordered, from most to less severe, all classifier functions
implemented for Andiacare.}
\usage{
andiacareClassifiers()
}
\value{
vector with ordered allowed class leveles
}
\description{
Returns, ordered, from most to less severe, all classifier functions
implemented for Andiacare.
}
\author{
Fernando Moreno Jabato <fmjabato(at)gmail(dot)com>
}
