% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_andiacare.R
\name{andiacareWrapper}
\alias{andiacareWrapper}
\title{Classifies a dataframe of patients using Andiacare algorithm}
\usage{
andiacareWrapper(df, cnames = NULL, type = "classify")
}
\arguments{
\item{df}{of patients to be classified}

\item{cnames}{(Optional) list with columnames relations}

\item{type}{(Optional) select between classify (default) or "extra"}
}
\value{
vector with patients classes
}
\description{
Classifies a dataframe of patients using Andiacare algorithm
}
\author{
Fernando Moreno Jabato <fmjabato(at)gmail(dot)com>
}
