% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{anonimate}
\alias{anonimate}
\title{Anonimates a dataframe given. Creates a new name for each patient (row) and
substitute it into dataframe using the names column and the surname column.
If surname column is not given (NULL), the software concatenate the full name
using the format <Surname; Name> and stores into names column.
If any error occurs, NULL will be returned}
\usage{
anonimate(df, nameCol, surnameCol = NULL, onlyDF = FALSE)
}
\arguments{
\item{df}{with cohort data}

\item{nameCol}{names column identifier or index}

\item{surnameCol}{(optional) surnames column identifier or index}

\item{onlyDF}{if TRUE, only the anonimated dataframe will be returned}
}
\value{
list with original, anonimated dictionary and anonimated df or
anonimated dataframe if onlyDF is setted to TRUE
}
\description{
Anonimates a dataframe given. Creates a new name for each patient (row) and
substitute it into dataframe using the names column and the surname column.
If surname column is not given (NULL), the software concatenate the full name
using the format <Surname; Name> and stores into names column.
If any error occurs, NULL will be returned
}
\author{
Fernando Moreno Jabato <fmjabato(at)gmail(dot)com>
}
