% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attd19_functions.R
\name{attd19CVTargetDM1}
\alias{attd19CVTargetDM1}
\title{Checks CV objective for ATTD2019 DM1 consensus of CV < 36\%}
\usage{
attd19CVTargetDM1(cv)
}
\arguments{
\item{cv}{vector of CV values}
}
\value{
TRUE if complains or FALSE in other cases
}
\description{
Checks CV objective for ATTD2019 DM1 consensus of CV < 36\%
}
\author{
Fernando Moreno Jabato <fmjabato(at)gmail(dot)com>
}
