% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attd19_functions.R
\name{attd19GlucoseTargetDM1}
\alias{attd19GlucoseTargetDM1}
\title{Checks if Glucose MEAN and SD complains ATTD2019 DM1 consesun of
Mean < 154 mg/dl and SD < 29\%}
\usage{
attd19GlucoseTargetDM1(means, sd)
}
\arguments{
\item{means}{vector of Mean Glucose (mg/dl)}

\item{sd}{vector of Standar Deviation of patients (\%)}
}
\value{
TRUE if complains or FALSE in other cases
}
\description{
Checks if Glucose MEAN and SD complains ATTD2019 DM1 consesun of
Mean < 154 mg/dl and SD < 29\%
}
\author{
Fernando Moreno Jabato <fmjabato(at)gmail(dot)com>
}
