% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attd19_functions.R
\name{attd19TARSTargetDM1}
\alias{attd19TARSTargetDM1}
\title{Checks TARS objective for ATTD19 consensus of TARS < 5\% for
DM1 patients}
\usage{
attd19TARSTargetDM1(tars)
}
\arguments{
\item{tars}{vector of TARS values}
}
\value{
TRUE if complains or FALSE in other cases
}
\description{
Checks TARS objective for ATTD19 consensus of TARS < 5\% for
DM1 patients
}
\author{
Fernando Moreno Jabato <fmjabato(at)gmail(dot)com>
}
