% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attd19_functions.R
\name{attd19TARTargetDM1}
\alias{attd19TARTargetDM1}
\title{Checks TAR objective for ATTD19 consensus of TAR < 25\% for
DM1 patients}
\usage{
attd19TARTargetDM1(tar)
}
\arguments{
\item{tar}{vector of TAR values}
}
\value{
TRUE if complains or FALSE in other cases
}
\description{
Checks TAR objective for ATTD19 consensus of TAR < 25\% for
DM1 patients
}
\author{
Fernando Moreno Jabato <fmjabato(at)gmail(dot)com>
}
