% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attd19_functions.R
\name{attd19TBRSTargetDM1}
\alias{attd19TBRSTargetDM1}
\title{Checks TBRS objective for ATTD19 consensus of TBRS <= 1\% for
DM1 patients}
\usage{
attd19TBRSTargetDM1(tbrs)
}
\arguments{
\item{tbrs}{vector of TBRS values}
}
\value{
TRUE if complains or FALSE in other cases
}
\description{
Checks TBRS objective for ATTD19 consensus of TBRS <= 1\% for
DM1 patients
}
\author{
Fernando Moreno Jabato <fmjabato(at)gmail(dot)com>
}
