% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attd19_functions.R
\name{attd19TBRTargetDM1}
\alias{attd19TBRTargetDM1}
\title{Checks TBR objective for ATTD19 consensus of TBR < 4\% for
DM1 patients}
\usage{
attd19TBRTargetDM1(tbr)
}
\arguments{
\item{tbr}{vector of TBR values}
}
\value{
TRUE if complains or FALSE in other cases
}
\description{
Checks TBR objective for ATTD19 consensus of TBR < 4\% for
DM1 patients
}
\author{
Fernando Moreno Jabato <fmjabato(at)gmail(dot)com>
}
