% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attd19_functions.R
\name{attd19TIRTargetDM1}
\alias{attd19TIRTargetDM1}
\title{Checks TIR objective for ATTD19 consensus of TIR > 70\% for
DM1 patients}
\usage{
attd19TIRTargetDM1(tir)
}
\arguments{
\item{tir}{vector of TIR values}
}
\value{
TRUE if complains or FALSE in other cases
}
\description{
Checks TIR objective for ATTD19 consensus of TIR > 70\% for
DM1 patients
}
\author{
Fernando Moreno Jabato <fmjabato(at)gmail(dot)com>
}
