% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_functions.R
\name{calculateFlowMatrix}
\alias{calculateFlowMatrix}
\title{Calculates the flow matrix for a given time series}
\usage{
calculateFlowMatrix(data, states, backward = 1, probability = FALSE)
}
\arguments{
\item{data}{dataframe with the time series}

\item{states}{vector with the possible states}

\item{backward}{number of previous states to be considered}

\item{probability}{if TRUE, the matrix is normalized to sum 1}
}
\value{
dataframe with the flow matrix
}
\description{
Calculates the flow matrix for a given time series
}
\author{
Fernando Moreno Jabato <fmjabato(at)gmail(dot)com>
}
