% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_gri.R
\name{calculatesMetaPolygonsGRI}
\alias{calculatesMetaPolygonsGRI}
\title{Internal function used to calculate area polygons for each class}
\usage{
calculatesMetaPolygonsGRI(GRImin, GRImax)
}
\arguments{
\item{GRImin}{minimum GRI of class range}

\item{GRImax}{maximum GRI of class range}
}
\value{
list with derised polygon corners
}
\description{
Internal function used to calculate area polygons for each class
}
\author{
Fernando Moreno Jabato <fmjabato(at)gmail(dot)com>
}
