% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_andiacare.R
\name{checkAndiacare}
\alias{checkAndiacare}
\title{Checks that necessary cols are available to perform classification}
\usage{
checkAndiacare(cohort, cnames = NULL)
}
\arguments{
\item{cohort}{dataframe to be checked}

\item{cnames}{(Optional) list with columnames relations}
}
\value{
TRUE if Andiacare algorithm can be applied or FALSE in other cases
}
\description{
Checks that necessary cols are available to perform classification
}
\author{
Fernando Moreno Jabato <fmjabato(at)gmail(dot)com>
}
