% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_functions.R
\name{checkEntryQA}
\alias{checkEntryQA}
\title{Checks a patient(s) entry in order to check if complies QA tests}
\usage{
checkEntryQA(data, algorithms = c("Andiacare", "GRI"), remove.ends = FALSE)
}
\arguments{
\item{data}{to be checked}

\item{algorithms}{that will be executed}

\item{remove.ends}{if TRUE, 0\% and 100\% will be removed}
}
\value{
TRUE if everything is ok or FALSE in other cases
}
\description{
Checks a patient(s) entry in order to check if complies QA tests
}
\author{
Fernando Moreno Jabato <fmjabato(at)gmail(dot)com>
}
