% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_functions.R
\name{classifyPatient}
\alias{classifyPatient}
\title{Classifies a patient using a given algorithm}
\usage{
classifyPatient(
  patient,
  algorithm = "Andiacare",
  cnames = NULL,
  type = "classify"
)
}
\arguments{
\item{patient}{dataframe with patient(s) to be classified}

\item{algorithm}{algorithm to be used}

\item{cnames}{(Optional) list with columnames needed to clasify. If its NULL,
default algorithm colnames will be used}

\item{type}{(Optional) select between classify (default) or (extra)}
}
\value{
vector with classes or NULL if classifier is not available
}
\description{
Classifies a patient using a given algorithm
}
\author{
Fernando Moreno Jabato <jabato(at)uma(dot)com
}
