% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{corrChart}
\alias{corrChart}
\title{Calculates correlation between columns of a cohort dataset}
\usage{
corrChart(
  df,
  targets = c("TIR", "TBR", "TBRS", "TAR", "TARS", "GMI", "MeanReads", "SourceSize",
    "SWEET", "CV", "Age", "AgeRange", "IllnessYears", "DebutAge"),
  method = "pearson",
  histogram = TRUE
)
}
\arguments{
\item{df}{with cohort data}

\item{targets}{columns that, if are available, will be studied}

\item{method}{correlation method to be applied}

\item{histogram}{plots histograms in main diagonal}
}
\value{
PerformanceAnalytics object with correlations
}
\description{
Calculates correlation between columns of a cohort dataset
}
\author{
Fernando Moreno Jabato <fmjabato(at)gmail(dot)com>
}
