% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_functions.R
\name{extractTS}
\alias{extractTS}
\title{Exrtacts a time series from a list of dataframes}
\usage{
extractTS(dataList, column)
}
\arguments{
\item{dataList}{list of dataframes}

\item{column}{column to be extracted}
}
\value{
dataframe with the time series
}
\description{
Exrtacts a time series from a list of dataframes
}
\author{
Fernando Moreno Jabato <fmjabato(at)gmail(dot)com>
}
