% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attd19_functions.R
\name{generalNumericVectorChecker}
\alias{generalNumericVectorChecker}
\title{Checks a lower/greater condition for a numeric vector}
\usage{
generalNumericVectorChecker(values, target, lower = TRUE, equals = FALSE)
}
\arguments{
\item{values}{vector to be checked}

\item{target}{to be checked}

\item{lower}{if TRUE, checks if values are lower than target}

\item{equals}{if TRUE, checks if values are X or equal than target}
}
\value{
TRUE if complains or FALSE in other cases
}
\description{
Checks a lower/greater condition for a numeric vector
}
\author{
Fernando Moreno Jabato <fmjabato(at)gmail(dot)com>
}
