% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates_functions.R
\name{getAge}
\alias{getAge}
\title{Calculates age from birthdate}
\usage{
getAge(birthdate, format = "\%d-\%m-\%Y")
}
\arguments{
\item{birthdate}{date}

\item{format}{(optional) date format}
}
\value{
age
}
\description{
Calculates age from birthdate
}
\author{
Fernando Moreno Jabato <fmjabato(at)gmail(dot)com>
}
