% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates_functions.R
\name{getDebutStats}
\alias{getDebutStats}
\title{Calculates debut related stats and return a dataframe with them}
\usage{
getDebutStats(df, debutCol = "Debut", format = "\%d-\%m-\%Y")
}
\arguments{
\item{df}{with Debut column}

\item{debutCol}{(optional) debut column name. Defualt: 'Debut'}

\item{format}{(optional) date format}

\item{birthdateCol}{(optional) birthdate column name. Defualt: 'Birthdate'}
}
\value{
dataframe with debut stats
}
\description{
Calculates debut related stats and return a dataframe with them
}
\author{
Fernando Moreno Jabato <fmjabato(at)gmail(dot)com>
}
