% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_gri.R
\name{gri}
\alias{gri}
\title{Classifies a dataframe of patients using GRI algorithm}
\usage{
gri(
  df,
  TIRcol = "TIR",
  TBRcol = "TBR",
  TARcol = "TAR",
  TBRScol = "TBRS",
  TARScol = "TARS"
)
}
\arguments{
\item{df}{of patients to be classified}

\item{TIRcol}{TIR info column name}

\item{TBRcol}{TBR info column name}

\item{TARcol}{TAR info column name}

\item{TBRScol}{TBRS info column name}

\item{TARScol}{TARS info column name}
}
\value{
vector with patients classes
}
\description{
Classifies a dataframe of patients using GRI algorithm
}
\author{
Fernando Moreno Jabato <fmjabato(at)gmail(dot)com>
}
