% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_gri.R
\name{griClasses}
\alias{griClasses}
\title{Returns, ordered, from most to less severe, all class levels implemented}
\usage{
griClasses(meta = FALSE)
}
\arguments{
\item{meta}{if TRUE, returns also specified metainfo per class}
}
\value{
vector with ordered allowed class leveles
}
\description{
Returns, ordered, from most to less severe, all class levels implemented
}
\author{
Fernando Moreno Jabato <fmjabato(at)gmail(dot)com>
}
