% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_gri.R
\name{griClassifiers}
\alias{griClassifiers}
\title{Returns, ordered, from most to less severe, all classifier functions
implemented for Glycemia Risk Index (GRI).
Note: doi:10.1177/19322968221085273}
\usage{
griClassifiers()
}
\value{
vector with ordered allowed class leveles
}
\description{
Returns, ordered, from most to less severe, all classifier functions
implemented for Glycemia Risk Index (GRI).
Note: doi:10.1177/19322968221085273
}
\author{
Fernando Moreno Jabato <fmjabato(at)gmail(dot)com>
}
