% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_functions.R
\name{hasExtra}
\alias{hasExtra}
\title{Checks if an algorithm can generate extra info}
\usage{
hasExtra(algorithm)
}
\arguments{
\item{algorithm}{to be checked}
}
\value{
TRUE if algorithm has extra or false in other cases
}
\description{
Checks if an algorithm can generate extra info
}
\author{
Fernando Moreno Jabato <fmjabato(at)gmail(dot)com>
}
