% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_functions.R
\name{inferTXR}
\alias{inferTXR}
\title{Correct missing TXR column if possible}
\usage{
inferTXR(df, tirCol = "TIR", tbrCol = "TBR", tarCol = "TAR")
}
\arguments{
\item{df}{data frame to be checked}

\item{tirCol}{(Optional) TIR colname}

\item{tbrCol}{(Optional) TBR colname}

\item{tarCol}{(Optional) TAR colname}
}
\value{
df corrected
}
\description{
Correct missing TXR column if possible
}
\author{
Fernando Moreno Jabato <fmjabato(at)gmail(dot)com>
}
