% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_andiacare.R
\name{isBlueAndiacare}
\alias{isBlueAndiacare}
\title{Checks if patients info given is classified as Blue flag.}
\usage{
isBlueAndiacare(info, readsCol = "MeanReads", thr = 4, logical = TRUE)
}
\arguments{
\item{info}{dataframe to be checked}

\item{readsCol}{mean read column name}

\item{thr}{threshold to marks as blue flag}

\item{logical}{if TRUE returns logical values, if FALSE returns labels}
}
\value{
logical vector, if TRUE, this patient is Blue flagged
}
\description{
Checks if patients info given is classified as Blue flag.
}
\author{
Fernando Moreno Jabato <fmjabato(at)gmail(dot)com>
}
