% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_andiacare.R
\name{isGreenAndiacare}
\alias{isGreenAndiacare}
\title{Checks if patients info given is classified as Green class.}
\usage{
isGreenAndiacare(info, meta = FALSE)
}
\arguments{
\item{info}{dataframe to be checked}

\item{meta}{if TRUE, add meta entries to result object}
}
\value{
logical vector, if TRUE, this patient is Green class
}
\description{
Checks if patients info given is classified as Green class.
}
\author{
Fernando Moreno Jabato <fmjabato(at)gmail(dot)com>
}
