% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_andiacare.R
\name{isRedAndiacare}
\alias{isRedAndiacare}
\title{Checks if patients info given is classified as Red class.}
\usage{
isRedAndiacare(info, meta = FALSE)
}
\arguments{
\item{info}{dataframe to be checked}

\item{meta}{if TRUE, add meta entry to result object}
}
\value{
logical vector, if TRUE, this patient is Red class
}
\description{
Checks if patients info given is classified as Red class.
}
\author{
Fernando Moreno Jabato <fmjabato(at)gmail(dot)com>
}
