% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libreview_functions.R
\name{libreviewCNAMES}
\alias{libreviewCNAMES}
\title{Gives libreview (Abbot) cnames set}
\usage{
libreviewCNAMES(lang = "ES")
}
\arguments{
\item{lang}{requested language}
}
\value{
libreview cnames requested or NULL if it's not supported
}
\description{
Gives libreview (Abbot) cnames set
}
\author{
Fernando Moreno Jabato <fmjabato(at)gmail(dot)com>
}
