% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libreview_functions.R
\name{loadLibreviewCSV}
\alias{loadLibreviewCSV}
\title{Loads a Libreview exported CSV file}
\usage{
loadLibreviewCSV(file, lang = NULL)
}
\arguments{
\item{file}{to be loaded}

\item{lang}{if not null, will be used to obtain libreview cnames}
}
\value{
a list with dataframe and metadata loaded
}
\description{
Loads a Libreview exported CSV file
}
\author{
Fernando Moreno Jabato <fmjabato(at)gmail(dot)com>
}
