% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{renderAndiacareReportDM1}
\alias{renderAndiacareReportDM1}
\title{Renders an andiacare cohort report}
\usage{
renderAndiacareReportDM1(
  cohort,
  output,
  template_folder = file.path(find.package("andiacare"), "templates"),
  verbose = FALSE,
  pediatric = FALSE,
  dateFormat = "\%d-\%m-\%Y"
)
}
\arguments{
\item{cohort}{to be plotted}

\item{output}{output file}

\item{template_folder}{(Optional) Rmd templates path}

\item{verbose}{(Optional) boolean that activates verbose mode}

\item{pediatric}{(Optional) boolean that activates pediatric mode}
}
\value{
void
}
\description{
Renders an andiacare cohort report
}
\author{
Fernando Moreno Jabato <fmjabato(at)gmail(dot)com>
}
