% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libreview_functions.R
\name{rl}
\alias{rl}
\title{Internal function used to read a Libreview CSV line and clean known issues}
\usage{
rl(connin, N = 1, csep = ",", NA.strings = c("NA", "*", "-", "", " "))
}
\arguments{
\item{connin}{input connection}

\item{N}{lines to be read}

\item{csep}{separator char}

\item{NA.strings}{strings to be read as NA}
}
\value{
line splitted
}
\description{
Internal function used to read a Libreview CSV line and clean known issues
}
\author{
Fernando Moreno Jabato <fmjabato(at)gmail(dot)com>
}
