% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweet_functions.R
\name{sizeRangesColorsSWEET}
\alias{sizeRangesColorsSWEET}
\title{Returns SWEET hospital size ranges colors vector}
\usage{
sizeRangesColorsSWEET()
}
\value{
SWEET size ranges colors vector
}
\description{
Returns SWEET hospital size ranges colors vector
}
\author{
Fernando Moreno Jabato <fmjabato(at)gmail(dot)com>
}
