% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweet_functions.R
\name{sizeRangeSWEET}
\alias{sizeRangeSWEET}
\title{Classify a column of hospitals using SWEET sizes ranges}
\usage{
sizeRangeSWEET(hospitals, factors = TRUE)
}
\arguments{
\item{hospitals}{vector to be checked}

\item{factors}{if TRUE, return vector of factors. Default: TRUE}
}
\value{
size ranges by SWEET
}
\description{
Classify a column of hospitals using SWEET sizes ranges
}
\author{
Fernando Moreno Jabato <fmjabato(at)gmail(dot)com>
}
