% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{statsTable}
\alias{statsTable}
\title{Creates a stats table for a cohort dataset and expected numeric tables}
\usage{
statsTable(
  dataset,
  decimals = 3,
  targets = c("TARS", "TAR", "TIR", "TBR", "TBRS", "MeanReads", "Age", "SourceSize",
    "DebutAge", "IllnessYears"),
  boolTargets = c(),
  toPlot = FALSE,
  title = ""
)
}
\arguments{
\item{dataset}{to be checked}

\item{decimals}{to round}

\item{targets}{to be included into table}

\item{boolTargets}{(optional) extra boolean to be included into table}

\item{toPlot}{activates to plot process and exports a tablegrob instead a
data frame.}

\item{title}{activates toPlot and includes given title}
}
\value{
stats dataframe (list if bool are included) or tableGrob object if
to plot is activated
}
\description{
Creates a stats table for a cohort dataset and expected numeric tables
}
\author{
Fernando Moreno Jabato <fmjabato(at)gmail(dot)com>
}
