% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tableFrom2Columns}
\alias{tableFrom2Columns}
\title{Generates a table from factor/character combinations of columns}
\usage{
tableFrom2Columns(df, columns, roundPercentage = 2, removeZeros = TRUE)
}
\arguments{
\item{df}{dataframe to be used}

\item{columns}{vector of indexes o columnames to be used}

\item{roundPercentage}{number of decimals to round percentage}

\item{removeZeros}{if TRUE, combinations with zero instances are removed}
}
\value{
table with amounts and percentages
}
\description{
Generates a table from factor/character combinations of columns
}
\author{
Fernando Moreno Jabato <fmjabato(at)gmail(dot)com>
}
