% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tableFromCharVector}
\alias{tableFromCharVector}
\title{Generates a table from factor/character vecto}
\usage{
tableFromCharVector(values)
}
\arguments{
\item{values}{vector of values/classes/labels}
}
\value{
table with amounts and percentages
}
\description{
Generates a table from factor/character vecto
}
\author{
Fernando Moreno Jabato <fmjabato(at)gmail(dot)com>
}
