% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{todayString}
\alias{todayString}
\title{Gives TODAY date as string}
\usage{
todayString()
}
\value{
today date as string
}
\description{
Gives TODAY date as string
}
\author{
Fernando Moreno Jabato <fmjabato(at)gmail(dot)com>
}
