% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_functions.R
\name{writeClassificationTable}
\alias{writeClassificationTable}
\title{Exports a cohort set adding metadata info}
\usage{
writeClassificationTable(df, outfile, sourceDate = NULL, analysisDate = NULL)
}
\arguments{
\item{df}{dataframe with classification}

\item{outfile}{output file}

\item{sourceDate}{string with original dataset date}

\item{analysisDate}{string with analysis date}
}
\value{
void
}
\description{
Exports a cohort set adding metadata info
}
\author{
Fernando Moreno Jabato <fmjabato(at)gmail(dot)com>
}
