% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libreview_functions.R
\name{writeLibreviewCSV}
\alias{writeLibreviewCSV}
\title{Writes data as Libreview exported CSV file}
\usage{
writeLibreviewCSV(
  df,
  outfile,
  metadata = NULL,
  header = NULL,
  originalLVFormat = FALSE
)
}
\arguments{
\item{df}{cohort dataset}

\item{outfile}{destiny}

\item{metadata}{(optional) pre-data header of file. If NULL, an empty entry
will be injected}

\item{header}{(optional) header colnames to be written}

\item{originalLVFormat}{if TRUE, original LibreView format will be exported.
If FALSE, corrected format will be used}
}
\value{
void
}
\description{
Writes data as Libreview exported CSV file
}
\author{
Fernando Moreno Jabato <fmjabato(at)gmail(dot)com>
}
